refactor(item_extractor): centralize all item-extraction logic
